перенес метод getFastestResultOfStream() из FacultyController в InfoController
