+++ src/main/java/ru/hogwarts/school/service/FacultyService.java
@@ -18,6 +18,4 @@ public interface FacultyService {
     Collection<Faculty> getAllByNameOrColor(String name, String color); // получить факультеты по имени или цвету
 
     String getLongestName(); //получить самое длинное имя факультета
-
-    Integer getFastestResultOfStream(); //получить результат вычисления стрима за наименьшее кол-во времени
 }
