+++ src/main/java/ru/hogwarts/school/service/InfoService.java
@@ -2,4 +2,6 @@ package ru.hogwarts.school.service;
 
 public interface InfoService {
     Integer getPort();
+
+    Integer getFastestResultOfStream(); //получить результат вычисления стрима за наименьшее кол-во времени
 }
