+++ src/main/java/ru/hogwarts/school/service/impl/FacultyServiceImpl.java
@@ -101,23 +101,6 @@ public class FacultyServiceImpl implements FacultyService {
                 .orElse("");
     }
 
-    @Override
-    public Integer getFastestResultOfStream() {
-        LOGGER.info("Was invoked method for get sum of iteration");
-        Long startTime = System.nanoTime();
-
-        int sum = 0;
-        for (int a = 1; a <= 1_000_000; a++) {
-            sum += a;
-        }
-
-        Long endTime = System.nanoTime();
-        long executionTime = endTime - startTime;
-
-        LOGGER.info("Sum was received successfully. Method completed in " + executionTime + "ns");
-        return sum; // использование цикла вместо Stream API более, чем в 4 раза ускоряет выполнение запроса
-    }
-
     private static void referenceNameMaker(Faculty faculty) {
         faculty.setName(capitalize(faculty.getName().toLowerCase()));
     }
